[fx] code refactor
